Display message in test play script.
